Add user-controllable filtration interval and speed entities
Convert read-only filtration interval sensors to time entities with
native time pickers, and convert speed label sensors to select entities
with dropdowns. This allows users to modify filtration schedules and
pump speeds directly from the HA UI.

- Add time.py with 6 time entities (interval 1-3 from/to)
- Move 3 timer speed entities from sensor to select platform
- Register Platform.TIME in __init__.py
- Update strings.json and icons.json for new entity types
