+++ custom_components/aquarite/icons.json
@@ -28,33 +28,6 @@
       "filtration_intel_time": {
         "default": "mdi:timer-outline"
       },
-      "filtration_interval_1_from": {
-        "default": "mdi:clock-start"
-      },
-      "filtration_interval_1_to": {
-        "default": "mdi:clock-end"
-      },
-      "filtration_interval_2_from": {
-        "default": "mdi:clock-start"
-      },
-      "filtration_interval_2_to": {
-        "default": "mdi:clock-end"
-      },
-      "filtration_interval_3_from": {
-        "default": "mdi:clock-start"
-      },
-      "filtration_interval_3_to": {
-        "default": "mdi:clock-end"
-      },
-      "filtration_timer_speed_1": {
-        "default": "mdi:speedometer"
-      },
-      "filtration_timer_speed_2": {
-        "default": "mdi:speedometer"
-      },
-      "filtration_timer_speed_3": {
-        "default": "mdi:speedometer"
-      },
       "pool_name": {
         "default": "mdi:pool"
       },
@@ -197,6 +170,35 @@
       },
       "pump_speed": {
         "default": "mdi:speedometer"
+      },
+      "filtration_timer_speed_1": {
+        "default": "mdi:speedometer"
+      },
+      "filtration_timer_speed_2": {
+        "default": "mdi:speedometer"
+      },
+      "filtration_timer_speed_3": {
+        "default": "mdi:speedometer"
+      }
+    },
+    "time": {
+      "filtration_interval_1_from": {
+        "default": "mdi:clock-start"
+      },
+      "filtration_interval_1_to": {
+        "default": "mdi:clock-end"
+      },
+      "filtration_interval_2_from": {
+        "default": "mdi:clock-start"
+      },
+      "filtration_interval_2_to": {
+        "default": "mdi:clock-end"
+      },
+      "filtration_interval_3_from": {
+        "default": "mdi:clock-start"
+      },
+      "filtration_interval_3_to": {
+        "default": "mdi:clock-end"
       }
     },
     "light": {
